TODO SUBTODO and screen connection

See merge request com6510-2023-24/team02/project!15

+++ app/src/main/java/com/team2/todo/data/daos/SubTodoDao.kt
@@ -1,11 +1,9 @@
 package com.team2.todo.data.daos
 
 import androidx.room.Dao
-import androidx.room.Insert
 import androidx.room.Query
 import androidx.room.Transaction
 import androidx.room.Upsert
-import com.team2.todo.data.entities.Images
 import com.team2.todo.data.entities.SubTodo
 import kotlinx.coroutines.flow.Flow
 
@@ -18,7 +16,7 @@ interface SubTodoDao {
     fun getSubTodosBasedOnTodo(todoId: Long): Flow<List<SubTodo>>
 
     @Query("SELECT  * FROM subtodos where subTodoId = :subTodoId ORDER BY priority ASC")
-    fun getSubTodosBasedOnSubTodoIdOrderedByPriority(subTodoId: Long): Flow<SubTodo>
+    fun getSubTodosBasedOnSubTodoIdOrderedByPriority(subTodoId: Int): Flow<SubTodo>
 
     @Transaction
     @Query("UPDATE subtodos SET status = :status WHERE subTodoId = :subTodoId")
